requirements.txt
# ========================================
# GeQie Monitor + AI Video Clip System
# Unified Python Dependencies
# ========================================

# ========================================
# Geqie Monitor Dependencies
# ========================================

# Web Framework (Flask)
Flask==3.1.2
Flask-Compress==1.23
Flask-SocketIO==5.4.1
python-socketio==5.11.4
python-engineio==4.13.0

# Flask Dependencies
Werkzeug==3.1.5
Jinja2==3.1.6
MarkupSafe==3.0.3

# Bilibili API
bilibili-api-python==17.4.1

# Data Processing
PyYAML==6.0.3
pytz==2025.2
apscheduler==3.11.2

# HTTP Clients
requests==2.32.5
aiohttp==3.13.3
aiofiles==23.2.1

# Other Dependencies
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
colorama==0.4.6

# Chinese NLP Libraries
snownlp==0.12.3
jieba==0.42.1
wordcloud==1.9.3

# Data Export Dependencies
openpyxl>=3.1.0
reportlab>=4.0.0

# ========================================
# AutoClip Dependencies
# ========================================

# Web Framework (FastAPI)
anyio>=4.7.0
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Data Validation
pydantic==2.11.0
pydantic-settings==2.6.0

# Environment Variables
python-dotenv==1.0.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Subtitle Processing
pysrt==1.1.2
webvtt-py==0.5.1

# Text Encoding
chardet==5.2.0

# Encryption
cryptography==41.0.7

# Image Processing
Pillow==10.1.0

# File Upload
python-multipart==0.0.12

# HTTP Clients
httpx==0.27.2

# Video Processing
ffmpeg-python>=0.2.0

# Logging
structlog>=24.1.0

# Retry Mechanism
tenacity>=8.2.3

# ========================================
# Notes
# ========================================
# 1. bcut-asr is included in backend/bcut-asr directory
#    Install with: pip install -e backend/bcut-asr
#
# 2. External Dependencies Required:
#    - Redis 5.0+ (for Celery task queue)
#    - FFmpeg 4.0+ (for video processing, add to system PATH)
#    - Node.js 18.x+ (for frontend)
#
# 3. Python Version: 3.9+
